fix: checks navigator language for rtl enabling
- RTL is now set from the configuration by priority. If no configuration
is defined for the rtl or language - navigator lanaguage is checked.
