+++ packages/base/src/util/detectNavigatorLanguage.js
@@ -1,0 +1,11 @@
+export default () => {
+	const browserLanguages = navigator.languages;
+
+	const navigatorLanguage = () => {
+		return navigator.language;
+	};
+
+	const rawLocale = (browserLanguages && browserLanguages[0]) || navigatorLanguage() || navigator.userLanguage || navigator.browserLanguage;
+
+	return rawLocale || "en";
+};
