+++ packages/base/src/util/getDesigntimePropertyAsArray.js
@@ -1,0 +1,4 @@
+export default value => {
+	const m = /\$([-a-z0-9A-Z._]+)(?::([^$]*))?\$/.exec(value);
+	return m && m[2] ? m[2].split(/,/) : null;
+};
